pms: Rename conversationsSelectors.
Part of a series replacing "conversation" as a jargon term
specific to PMs, with "PM conversation".
